+++ R/jsonStrings.R
@@ -1,307 +1,458 @@
 #' @useDynLib jsonStrings, .registration=TRUE
-#' @importFrom Rcpp evalCpp  
+#' @importFrom Rcpp evalCpp setRcppClass 
 #' @importFrom methods new
 NULL
 
-Rcpp::loadModule("jsonModule", what = "JSON")
-Rcpp::loadModule("jsonptrModule", what = "JSONPTR")
+JsonString <- setRcppClass("JsonString")
 
-#' @title JSON string
-#' @description Create a JSON string.
-#'
-#' @param string a character string
-#'
-#' @return A JSON string (external pointer).
-#' @export
-#'
-#' @examples jsonString("[1,[\"a\",99],{\"x\":[2,3,4],\"y\":42}]")
-jsonString <- function(string){
-  new(JSON, string)$jsonPointer()
+Xptr <- function(jstring){
+  jstring[[".__enclos_env__"]][["private"]][[".jsonString"]][["ptr"]]
 }
 
-#' @title Access an element in a JSON string
-#' @description Extract an element in a JSON string by giving a path of keys or 
-#'   indices.
+Xptrinit <- function(ptr){
+  json <- jsonString$new("{}")
+  json[[".__enclos_env__"]][["private"]][[".jsonString"]] <- 
+    new("JsonString", ptr, 0L)
+  json
+}
+
+
+#' @title R6 class to represent a JSON string
+#' @description R6 class to represent a JSON string.
 #'
-#' @param jsonstring a JSON string representing an array or an object
-#' @param path path to the element to extract; either a character vector of 
-#'   keys, an integer vector of indices, or a list made of keys and indices
-#'
-#' @return A JSON string.
-#' @export
-#'
-#' @examples jstring <- "[1,[\"a\",99],{\"x\":[2,3,4],\"y\":42}]"
-#' jsonAt(jstring, 1)
-#' jsonAt(jstring, list(2, "x"))
-#' jsonAt(jstring, list(2, "z"))
-jsonAt <- function(jsonstring, path){
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  if(is.list(path)){
-    if(!checkPath(path)){
-      stop("Invalid path.", call. = TRUE)
+#' @importFrom R6 R6Class
+#' @export 
+jsonString <- R6Class(
+  "jsonString",
+  
+  lock_class = TRUE,
+  
+  cloneable = FALSE,
+  
+  private = list(
+    .prettyPrint = TRUE,
+    .jsonString = NULL
+  ),
+  
+  active = list(
+    #' @field prettyPrint get or set the value of \code{prettyPrint}
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "[1, [\"a\", 99], {\"x\": [2,3,4], \"y\": 42}]"
+    #' )
+    #' jstring$prettyPrint
+    #' jstring
+    #' jstring$prettyPrint <- FALSE
+    #' jstring
+    prettyPrint = function(value) {
+      if(missing(value)) {
+        private[[".prettyPrint"]]
+      } else {
+        stopifnot(
+          isBoolean(value)
+        )
+        private[[".prettyPrint"]] <- value
+      }
     }
-    keys <- unlist(path)
-    indices <- suppressWarnings(as.integer(keys))
-    isIndex <- vapply(path, is.numeric, FUN.VALUE = logical(1L))
-  }else if(is.numeric(path)){
-    keys <- ""
-    indices <- as.integer(path)
-    isIndex <- rep(TRUE, length(path))
-  }else if(is.character(path)){
-    keys <- path
-    indices <- 0L
-    isIndex <- rep(FALSE, length(path))
-  }else{
-    stop("Invalid path.", call. = TRUE)
-  }
-  if(any(indices < 0L, na.rm = TRUE)){
-    stop("Negative indices found in path.", call. = TRUE)
-  }
-  new(JSONPTR, jsonstring)$at(keys = keys, indices = indices, isIndex = isIndex)
-}
+  ),
+  
+  public = list(
+    
+    #' @description Creates a new \code{jsonString} object.
+    #'
+    #' @param string a string representing a JSON object, or the path to a 
+    #'   JSON file
+    #'
+    #' @return A \code{jsonString} object.
+    #'
+    #' @examples
+    #' jstring <- "[1, [\"a\", 99], {\"x\": [2,3,4], \"y\": 42}]"
+    #' jsonString$new(jstring)
+    initialize = function(string){
+      stopifnot(isString(string))
+      if(file.exists(string)){
+        ptr <- read_json(string)
+        private[[".jsonString"]] <- JsonString$new(ptr, 0L)
+      }else{
+        private[[".jsonString"]] <- JsonString$new(string)
+      }
+      invisible(self)
+    },
+    
+    #' @description Print a \code{jsonString} object.
+    #' @param ... ignored
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "[1, [\"a\", 99], {\"x\": [2,3,4], \"y\": 42}]"
+    #' )
+    #' jstring
+    #' jstring$prettyPrint <- FALSE
+    #' jstring
+    print = function(...){
+      private[[".jsonString"]]$print(pretty = private[[".prettyPrint"]])
+    },
+    
+    #' @description Converts a \code{jsonString} to a character string.
+    #' @param pretty Boolean, whether to get a pretty string
+    #' @return A string.
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "[1, [\"a\", 99], {\"x\": [2,3,4], \"y\": 42}]"
+    #' )
+    #' cat(jstring$asString())
+    #' cat(jstring$asString(pretty = TRUE))
+    asString = function(pretty = FALSE){
+      stopifnot(isBoolean(pretty))
+      private[[".jsonString"]]$asString(pretty)
+    },
 
+    #' @description Extract an element in a JSON string by giving a path of 
+    #'   keys or indices.
+    #' @param ... the elements forming the path, integers or strings
+    #' @return A \code{jsonString} object.
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "[1, [\"a\", 99], {\"x\": [2,3,4], \"y\": 42}]"
+    #' )
+    #' jstring$at(1)
+    #' jstring$at(2, "x")
+    at = function(...){
+      ptr <- private[[".jsonString"]]$at(list(...))
+      Xptrinit(ptr)
+    },
+    
+    #' @description Checks whether a key exists in the reference JSON string.
+    #' @param key a string
+    #' @return A Boolean value.
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "[1, [\"a\", 99], {\"x\": [2,3,4], \"y\": 42}]"
+    #' )
+    #' jstring$hasKey("x")
+    #' jstring <- jsonString$new(
+    #'   "{\"x\": [2,3,4], \"y\": 42}"
+    #' )
+    #' jstring$hasKey("x")
+    hasKey = function(key){
+      stopifnot(isString(key))
+      private[[".jsonString"]]$hasKey(key)
+    },
 
-#' @title Does key exist?
-#' @description Checks whether a key is present in a JSON string.
-#'
-#' @param jsonstring a JSON string
-#' @param key character string
-#'
-#' @return \code{TRUE} if the given key is present in the JSON string, 
-#'   \code{FALSE} otherwise.
-#' @export
-#'
-#' @examples jsonHasKey("{\"a\":[1,2,3],\"b\":\"hello\"}", "b")
-#' jsonHasKey("[1,2,3]", "a")
-jsonHasKey <- function(jsonstring, key){
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  new(JSONPTR, jsonstring)$hasKey(key)
-}
+    #' @description Get the keys of the reference JSON string (if it represents 
+    #'   an object).
+    #' @return A character vector.
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"x\": [2,3,4], \"y\": 42}"
+    #' )
+    #' jstring$keys()
+    keys = function(){
+      private[[".jsonString"]]$keys()
+    },
+    
+    #' @description Add a new property to the reference JSON string (if it 
+    #'   represents an object).
+    #'
+    #' @param key a character string, the key of the new property
+    #' @param value a JSON string, either a \code{jsonString} object or a 
+    #'   string 
+    #'
+    #' @return This updates the reference JSON string and this returns it 
+    #'  invisibly.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' ppty <- jsonString$new("[9, 99]")
+    #' jstring$addProperty("c", ppty)
+    #' jstring
+    #' jstring$addProperty("d", "null")
+    #' jstring
+    addProperty = function(key, value){
+      stopifnot(isString(key))
+      if(isJsonString(value)){
+        ptr <- Xptr(value)
+      }else if(isString(value)){
+        ptr <- toJSONXptr(value)
+      }else{
+        stop("Invalid `value` argument.")
+      }
+      private[[".jsonString"]]$addProperty(key, ptr)
+      invisible(self)
+    },
+    
+    #' @description Erase an object property or an array element from the 
+    #'   reference JSON string.
+    #'
+    #' @param at either a character string, the key of the property to be erased, 
+    #'   or an integer, the index of the array element to be erased
+    #'
+    #' @return This invisibly returns the updated reference JSON string.
+    #'
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jstring$erase("b")
+    #' jstring
+    #' jstring <- jsonString$new("[1, 2, 3, 4, 5]")
+    #' jstring$erase(2)
+    #' jstring
+    erase = function(at){
+      if(isString(at)){
+        private[[".jsonString"]]$eraseProperty(at)
+        invisible(self)
+      }else if(isPositiveInteger(at)){
+        private[[".jsonString"]]$eraseElement(as.integer(at))
+        invisible(self)
+      }else{
+        stop("Invalid `at` argument.")
+      }
+    },
+    
+    #' @description Number of elements in the reference JSON string.
+    #' 
+    #' @return An integer.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jstring$size()
+    size = function(){
+      private[[".jsonString"]]$size()
+    },
+    
+    #' @description Update the reference JSON string (if it 
+    #'   represents an object).
+    #'
+    #' @param jstring a JSON string representing an object, either a 
+    #'   \code{jsonString} object or a string 
+    #'
+    #' @return This invisibly returns the updated reference JSON string.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jstring2 <- "{\"a\":[4,5,6],\"c\":\"goodbye\"}"
+    #' jstring$update(jstring2)
+    #' jstring
+    update = function(jstring){
+      if(isJsonString(jstring)){
+        ptr <- Xptr(jstring)
+      }else if(isString(jstring)){
+        ptr <- toJSONXptr(jstring)
+      }else{
+        stop("Invalid `jstring` argument.")
+      }
+      private[[".jsonString"]]$update(ptr)
+      invisible(self)
+    },
 
-#' @title Add new property
-#' @description Add a new property to a JSON string.
-#'
-#' @param jsonstring a JSON string representing an object
-#' @param key a character string, the key of the new property
-#' @param value a JSON string, the value of the new property
-#'
-#' @return A JSON string.
-#' @export
-#'
-#' @examples jsonAddProperty("{\"a\":[1,2,3],\"b\":\"hello\"}", "c", "[1,2]")
-jsonAddProperty <- function(jsonstring, key, value){
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  if(is.character(value)){
-    value <- jsonString(value)
-  }
-  new(JSONPTR, jsonstring)$addProperty(key, value)
-}
+    #' @description Merge the reference JSON string (if it 
+    #'   represents an object).
+    #'
+    #' @param jstring a JSON string, either a \code{jsonString} object or a 
+    #'   string representing a JSON object
+    #'
+    #' @return This invisibly returns the updated reference JSON string.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jstring2 <- "{\"a\":[4,5,6],\"c\":\"goodbye\"}"
+    #' jstring$merge(jstring2)
+    #' jstring
+    merge = function(jstring){
+      if(isJsonString(jstring)){
+        ptr <- Xptr(jstring)
+      }else if(isString(jstring)){
+        ptr <- toJSONXptr(jstring)
+      }else{
+        stop("Invalid `jstring` argument.")
+      }
+      private[[".jsonString"]]$merge(ptr)
+      invisible(self)
+    },
+    
+    
+    #' @description Apply a JSON patch to the reference JSON string (if it 
+    #'   represents an array or an object).
+    #'
+    #' @param jspatch a JSON patch, a JSON string representing an array (see 
+    #'   the link in details); it could be either a \code{jsonString} object or 
+    #'   a string 
+    #'
+    #' @return A new \code{jsonString} object.
+    #' 
+    #' @details See \href{http://jsonpatch.com/}{jsonpatch.com}.
+    #'
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jspatch <- "[
+    #'   {\"op\": \"remove\", \"path\": \"/a\"},
+    #'   {\"op\": \"replace\", \"path\": \"/b\", \"value\": null}
+    #' ]"
+    #' jstring$patch(jspatch)
+    patch = function(jspatch){
+      if(isJsonString(jspatch)){
+        ptrpatch <- Xptr(jspatch)
+      }else if(isString(jspatch)){
+        ptrpatch <- toJSONXptr(jspatch)
+      }else{
+        stop("Invalid `jspatch` argument.")
+      }
+      ptr <- private[[".jsonString"]]$patch(ptrpatch)
+      Xptrinit(ptr)
+    },
+    
+    #' @description Append an element to the reference JSON string (if it 
+    #'   represents an array).
+    #'
+    #' @param jstring a JSON string, either a \code{jsonString} object or a 
+    #'   string representing a JSON object
+    #'
+    #' @return This invisibly returns the updated reference JSON string.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new("[1, 2, 3, 4, 5]")
+    #' jstring2 <- jsonString$new(
+    #'   "{\"a\":[4,5,6],\"c\":\"goodbye\"}"
+    #'  )
+    #' jstring$push(jstring2)
+    #' jstring
+    push = function(jstring){
+      if(isJsonString(jstring)){
+        ptr <- Xptr(jstring)
+      }else if(isString(jstring)){
+        ptr <- toJSONXptr(jstring)
+      }else{
+        stop("Invalid `jstring` argument.")
+      }
+      private[[".jsonString"]]$push(ptr)
+      invisible(self)
+    },
+    
+    #' @description Check the type of the reference JSON string.
+    #'
+    #' @param type the type to be checked, one of \code{"array"}, \code{"object"}, 
+    #'   \code{"string"}, \code{"number"}, \code{"integer"}, \code{"float"}, 
+    #'   \code{"null"}, \code{"boolean"}
+    #'
+    #' @return A Boolean value.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jstring$is("object")
+    #' jstring$is("array")
+    #' jstring <- jsonString$new("999")
+    #' jstring$is("integer")
+    #' jstring$is("number")
+    #' jstring$is("float")
+    is = function(type){
+      types <-
+        c("array",
+          "object",
+          "number",
+          "integer",
+          "string",
+          "float",
+          "null",
+          "boolean")
+      type <- match.arg(type, types)
+      private[[".jsonString"]]$is(type) 
+    },
+    
+    #' @description Get the type of the reference JSON string.
+    #'
+    #' @return A character string indicating the type of the JSON string.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jstring$type()
+    #' jstring <- jsonString$new("999")
+    #' jstring$type()
+    type = function(){
+      private[[".jsonString"]]$type() 
+    },
+    
+    #' @description Flatten the reference JSON string.
+    #'
+    #' @return A new \code{jsonString} object.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":{\"x\":\"hello\",\"y\":\"hi\"}}"
+    #' )
+    #' jstring$flatten()
+    flatten = function(){
+      ptr <- private[[".jsonString"]]$flatten()
+      Xptrinit(ptr)
+    },
 
-#' @title Erase property or element
-#' @description Erase an object property or an array element from a JSON string.
-#'
-#' @param jsonstring a JSON string representing an object or an array
-#' @param at either a character string, the key of the property to be erased, 
-#'   or an integer, the index of the array element to be erased
-#'
-#' @return A JSON string.
-#' @export
-#'
-#' @examples jsonErase("{\"a\":[1,2,3],\"b\":\"hello\"}", "a")
-jsonErase <- function(jsonstring, at){
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  if(is.character(at)){
-    new(JSONPTR, jsonstring)$eraseProperty(at)
-  }else if(is.numeric(at) && at >= 0){
-    new(JSONPTR, jsonstring)$eraseElement(as.integer(at))
-  }else{
-    stop("Invalid `at` argument", call. = TRUE)
-  }
-}
-
-#' @title Number of elements
-#' @description Number of elements in a JSON string.
-#'
-#' @param jsonstring a JSON string 
-#'
-#' @return An integer.
-#' @export
-#'
-#' @examples jsonSize("{\"a\":[1,2,3],\"b\":\"hello\"}")
-jsonSize <- function(jsonstring){
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  new(JSONPTR, jsonstring)$size()
-}
-
-#' @title Update object
-#' @description Update a JSON string.
-#'
-#' @param jsonstring1 a JSON string representing an object
-#' @param jsonstring2 a JSON string representing an object
-#'
-#' @return A JSON string.
-#' @export
-#'
-#' @examples jstring1 <- jsonString("{\"a\":[1,2,3],\"b\":\"hello\"}")
-#' jstring2 <- jsonString("{\"a\":[4,5,6],\"c\":\"goodbye\"}")
-#' jsonUpdate(jstring1, jstring2)
-jsonUpdate <- function(jsonstring1, jsonstring2){
-  if(is.character(jsonstring1)){
-    jsonstring1 <- jsonString(jsonstring1)
-  }
-  if(is.character(jsonstring2)){
-    jsonstring2 <- jsonString(jsonstring2)
-  }
-  new(JSONPTR, jsonstring1)$update(jsonstring2)
-}
-
-#' @title Patch a JSON string
-#' @description Apply a JSON patch to a JSON string.
-#'
-#' @param jsondoc a JSON string representing an object or an array
-#' @param jsonpatch a JSON patch, a JSON string representing an array (see 
-#'   the link in details)
-#'
-#' @return A JSON string.
-#' @export
-#' 
-#' @details See \href{http://jsonpatch.com/}{jsonpatch.com}.
-#'
-#' @examples jsondoc <- jsonString("{\"a\":[1,2,3],\"b\":\"hello\"}")
-#' jsonpatch <- jsonString("[
-#'   {\"op\": \"remove\", \"path\": \"/a\"},
-#'   {\"op\": \"replace\", \"path\": \"/b\", \"value\": null}
-#' ]")
-#' jsonPatch(jsondoc, jsonpatch)
-jsonPatch <- function(jsondoc, jsonpatch){
-  if(is.character(jsondoc)){
-    jsondoc <- jsonString(jsondoc)
-  }
-  if(is.character(jsonpatch)){
-    jsonpatch <- jsonString(jsonpatch)
-  }
-  new(JSONPTR, jsondoc)$patch(jsonpatch)
-}
-
-#' @title Merge JSON strings
-#' @description Merge two JSON strings.
-#'
-#' @param jsonstring1 a JSON string 
-#' @param jsonstring2 a JSON string 
-#'
-#' @return A JSON string.
-#' @export
-#'
-#' @examples jstring1 <- jsonString("{\"a\":[1,2,3],\"b\":\"hello\"}")
-#' jstring2 <- jsonString("{\"a\":[4,5,6],\"c\":\"goodbye\"}")
-#' jsonMerge(jstring1, jstring2)
-jsonMerge <- function(jsonstring1, jsonstring2){
-  if(is.character(jsonstring1)){
-    jsonstring1 <- jsonString(jsonstring1)
-  }
-  if(is.character(jsonstring2)){
-    jsonstring2 <- jsonString(jsonstring2)
-  }
-  new(JSONPTR, jsonstring1)$merge(jsonstring2)
-}
-
-#' @title Append an element
-#' @description Append an element to a JSON string.
-#'
-#' @param jsonstring1 JSON string representing an array
-#' @param jsonstring2 JSON string 
-#'
-#' @return A JSON string.
-#' @export
-#'
-#' @examples jstring1 <- jsonString("[1,2,3]")
-#' jstring2 <- jsonString("{\"a\":\"hello\",\"b\":\"goodbye\"}")
-#' jsonPush(jstring1, jstring2)
-jsonPush <- function(jsonstring1, jsonstring2){
-  if(is.character(jsonstring1)){
-    jsonstring1 <- jsonString(jsonstring1)
-  }
-  if(is.character(jsonstring2)){
-    jsonstring2 <- jsonString(jsonstring2)
-  }
-  new(JSONPTR, jsonstring1)$push(jsonstring2)
-}
-
-#' @title Check type of JSON string 
-#' @description Check the type of a JSON string.
-#'
-#' @param jsonstring a JSON string
-#' @param type the type to be checked, one of \code{"array"}, \code{"object"}, 
-#'   \code{"string"}, \code{"number"}, \code{"integer"}, \code{"null"}, 
-#'   \code{"boolean"}
-#'
-#' @return A logical value.
-#' @export
-#' 
-#' @examples jsonIs("[1,2]", "array")
-jsonIs <- function(jsonstring, type){
-  types <- 
-    c("array", "object", "number", "integer", "string", "null", "boolean")
-  type <- match(match.arg(type, types), types)
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  new(JSONPTR, jsonstring)$is(type)
-}
-
-#' @title Type of JSON string 
-#' @description The type of a JSON string.
-#'
-#' @param jsonstring a JSON string
-#'
-#' @return A character string indicating the type of the JSON string.
-#' @export
-#' 
-#' @examples jsonType("[1,2]")
-jsonType <- function(jsonstring){
-  if(is.character(jsonstring)){
-    jsonstring <- jsonString(jsonstring)
-  }
-  new(JSONPTR, jsonstring)$type()
-}
-
-#' @title JSON string to character string
-#' @description Convert a JSON string to a character string.
-#'
-#' @param x a JSON string
-#' @param pretty logical value, whether to pretty-format the string
-#' @param ... ignored
-#' 
-#' @return A character string.
-#' @export
-as.character.jsonString <- function(x, pretty = FALSE, ...){
-  new(JSONPTR, x)$jsonString(pretty = pretty)
-}
-
-#' @title Print JSON string
-#' @description Print a JSON string.
-#'
-#' @param x a JSON string
-#' @param pretty logical value, whether to pretty-print
-#' @param ... ignored
-#'
-#' @export
-print.jsonString <- function(
-  x, pretty = getOption("jsonStrings.prettyPrint", FALSE), ...
-){
-  if(pretty){
-    cat(as.character(x, pretty = TRUE))
-  }else{
-    print(as.character(x))
-  }
-}
+    #' @description Unflatten the reference JSON string (if it is flattened).
+    #'
+    #' @return A new \code{jsonString} object.
+    #' 
+    #' @examples 
+    #' folder <- system.file(package = "jsonStrings")
+    #' files <- list.files(folder, recursive = TRUE)
+    #' sizes <- file.size(file.path(folder, files))
+    #' files <- sprintf('"%s"', paste0("/", files))
+    #' string <- sprintf("{%s}", paste0(files, ":", sizes, collapse = ","))
+    #' jstring <- jsonString$new(string)
+    #' jstring$unflatten()
+    unflatten = function(){
+      ptr <- private[[".jsonString"]]$unflatten()
+      Xptrinit(ptr)
+    },
+    
+    #' @description Write the reference JSON string to a file.
+    #' 
+    #' @param filename name of the file
+    #'
+    #' @return Nothing.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' jsonfile <- tempfile(fileext = ".json")
+    #' jstring$writeFile(jsonfile)
+    #' cat(readLines(jsonfile), sep = "\n")
+    #' jsonString$new(jsonfile)
+    writeFile = function(filename){
+      stopifnot(isString(filename))
+      private[[".jsonString"]]$writeFile(filename)
+    },
+    
+    #' @description Copy the reference JSON string.
+    #' 
+    #' @return A new \code{jsonString} object.
+    #' 
+    #' @examples 
+    #' jstring <- jsonString$new(
+    #'   "{\"a\":[1,2,3],\"b\":\"hello\"}"
+    #' )
+    #' copy <- jstring$copy()
+    #' copy$erase("b")
+    #' jstring
+    #' naive_copy <- jstring
+    #' naive_copy$erase("b")
+    #' jstring
+    copy = function(){
+      Xptrinit(Xptr(self))
+    }
+    
+  )
+  
+)