+++ NEWS.md
@@ -1,3 +1,12 @@
+# jsonStrings 2.0.0
+
+- Major change: introduced the R6 class `jsonString`.
+
+- A `jsonString` can be initialized from a JSON file.
+
+- Possibility to write a JSON string to a file (method `writeFile`).
+
+
 # jsonStrings 1.0.0
 
-* First release.
+First release.